+++ src/main/kotlin/com/github/surpsg/diffcoverage/properties/gradle/DiffCoverageGradlePlugin.kt
@@ -1,8 +1,5 @@
 package com.github.surpsg.diffcoverage.properties.gradle
 
 const val DIFF_COVERAGE_CONFIGURATION_PROPERTY = "diffCoverageReport"
-const val DIFF_COVERAGE_TASK = "diffCoverage"
 
 const val DIFF_COVERAGE_FAKE_TASK = "projects"
-
-const val HTML_REPORT_RELATIVE_PATH = "diffCoverage/html/index.html"