+++ src/main/kotlin/com/github/surpsg/diffcoverage/runconfiguration/DiffCoverageRunProfileState.kt
@@ -4,11 +4,7 @@ import com.github.surpsg.diffcoverage.DiffCoverageBundle
 import com.github.surpsg.diffcoverage.coroutine.BACKGROUND_SCOPE
 import com.github.surpsg.diffcoverage.domain.CoverageStat
 import com.github.surpsg.diffcoverage.domain.DiffCoverageConfiguration
-import com.github.surpsg.diffcoverage.domain.gradle.GradleModule
 import com.github.surpsg.diffcoverage.properties.NOT_GRADLE_PROJECT
-import com.github.surpsg.diffcoverage.properties.REPORT_LINK
-import com.github.surpsg.diffcoverage.properties.REPORT_LINK_WITH_ERROR
-import com.github.surpsg.diffcoverage.properties.gradle.HTML_REPORT_RELATIVE_PATH
 import com.github.surpsg.diffcoverage.services.CoverageVizualizeService
 import com.github.surpsg.diffcoverage.services.gradle.GradleDiffCoveragePluginSettingsService
 import com.github.surpsg.diffcoverage.services.gradle.GradleDiffCoverageRunService
@@ -18,13 +14,11 @@ import com.intellij.execution.ExecutionResult
 import com.intellij.execution.Executor
 import com.intellij.execution.configurations.RunProfileState
 import com.intellij.execution.runners.ProgramRunner
-import com.intellij.notification.NotificationListener
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import kotlinx.coroutines.launch
 import org.jacoco.core.analysis.ICoverageNode
-import java.nio.file.Paths
 
 class DiffCoverageRunProfileState(
     private val project: Project,
@@ -47,72 +41,61 @@ class DiffCoverageRunProfileState(
         val module = project.service<GradleDiffCoveragePluginSettingsService>().lookupDiffCoveragePluginModule()
             ?: return@launch
         project.service<GradleDiffCoverageRunService>().obtainCacheableDiffCoverageInfo(module)?.let { coverageInfo ->
-            launch {
-                buildCoverageReportByGradlePlugin(module, coverageInfo)
-            }
-            buildReportIde(coverageInfo)
+            showCoverageReportNotification(coverageInfo)
         }
     }
 
-    private fun buildReportIde(coverageInfo: DiffCoverageConfiguration) {
-        if (!diffCoverageRunConfiguration.buildReportByIde) {
-            return
-        }
+    private fun showCoverageReportNotification(coverageInfo: DiffCoverageConfiguration) {
         val coverageStat: CoverageStat = project.service<CoverageVizualizeService>().showCoverage(coverageInfo)
             ?: return
 
         val coverageHasViolations = getCoverageViolations(coverageStat).values.contains(true)
+        val (notificationType, title) = getIconToMessageKey(coverageHasViolations)
         project.service<BalloonNotificationService>().notify(
-            title = DiffCoverageBundle.message("coverage.stat.success.title"),
-            notificationType = if (coverageHasViolations) NotificationType.ERROR else NotificationType.INFORMATION,
+            title = title,
+            notificationType = notificationType,
             message = DiffCoverageBundle.message(
                 "coverage.stat.success.message",
-                coverageStat.linesCoverage,
-                coverageStat.branchesCoverage,
-                coverageStat.instructionsCoverage
+                coverageStat.linesCoverage.asPercentsText(),
+                coverageStat.branchesCoverage.asPercentsText(),
+                coverageStat.instructionsCoverage.asPercentsText()
             )
         )
     }
 
+    private fun getIconToMessageKey(coverageHasViolations: Boolean) = if (coverageHasViolations) {
+        NotificationType.ERROR to DiffCoverageBundle.message(
+            "coverage.stat.fail.title",
+            diffCoverageRunConfiguration.minCoveragePercents
+        )
+    } else {
+        NotificationType.INFORMATION to DiffCoverageBundle.message("coverage.stat.success.title")
+    }
+
     private fun getCoverageViolations(coverageStat: CoverageStat): Map<ICoverageNode.CounterEntity, Boolean> {
         return sequenceOf(
-            ICoverageNode.CounterEntity.LINE to MIN_COVERAGE_RATIO,
-            ICoverageNode.CounterEntity.BRANCH to MIN_COVERAGE_RATIO,
-            ICoverageNode.CounterEntity.INSTRUCTION to MIN_COVERAGE_RATIO,
+            ICoverageNode.CounterEntity.LINE,
+            ICoverageNode.CounterEntity.BRANCH,
+            ICoverageNode.CounterEntity.INSTRUCTION,
         ).map {
-            it.first to (coverageStat.getCoverage(it.first).coveredRatio < it.second)
+            it to coverageStat.getCoverage(it).coveredRatio
+        }.filter { (_, coverageRation) ->
+            !coverageRation.isNaN()
+        }.map { (coverageType, coverageRatio) ->
+            coverageType to (coverageRatio.toPercents() < diffCoverageRunConfiguration.minCoveragePercents)
         }.toMap()
     }
 
-    private suspend fun buildCoverageReportByGradlePlugin(
-        gradleModule: GradleModule,
-        coverageInfo: DiffCoverageConfiguration
-    ) {
-        if (!diffCoverageRunConfiguration.buildReportByGradlePlugin) {
-            return
-        }
-        val successExecute = project.service<GradleDiffCoverageRunService>().runDiffCoverageTask(gradleModule)
-        val messageKeyToNotificationType = if (successExecute) {
-            REPORT_LINK to NotificationType.INFORMATION
-        } else {
-            REPORT_LINK_WITH_ERROR to NotificationType.ERROR
-        }
-        showDiffCoverageReportNotification(coverageInfo, messageKeyToNotificationType)
-    }
+    private fun Double.toPercents(): Int = this.toInt() * TO_PERCENT_MULTIPLIER
 
-    private fun showDiffCoverageReportNotification(
-        diffCoverageInfo: DiffCoverageConfiguration,
-        messageKeyToNotificationType: Pair<String, NotificationType>
-    ) {
-        val reportUrl = Paths.get(diffCoverageInfo.reportsRoot, HTML_REPORT_RELATIVE_PATH).toUri().toString()
-        project.service<BalloonNotificationService>().notify(
-            notificationType = messageKeyToNotificationType.second,
-            notificationListener = NotificationListener.URL_OPENING_LISTENER,
-            message = DiffCoverageBundle.message(messageKeyToNotificationType.first, reportUrl)
-        )
+    private fun Double.asPercentsText(): String = if (isNaN()) {
+        NOT_AVAILABLE_COVERAGE
+    } else {
+        "$this%"
     }
 
     private companion object {
-        const val MIN_COVERAGE_RATIO = 0.9
+        const val TO_PERCENT_MULTIPLIER = 100
+        const val NOT_AVAILABLE_COVERAGE = "N/A"
     }
 }